% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Title}
\usage{
cluster(dt_val)
}
\arguments{
\item{a}{data frame to be added cluster label}
}
\value{
a data frame with cluster label
}
\description{
Title
}
\examples{
clustered <- cluster(validation_data)
}
